+++ tests/headers/func_no_proto.h
@@ -1,0 +1,2 @@
+// http://clang.llvm.org/doxygen/classclang_1_1FunctionNoProtoType.html
+typedef int no_proto();